fix: HttpArena json/upload validation failures
- Remove /json from pre-encoded cache so dataset handler runs
- Merge TechEmpower and HttpArena /json handlers (falls back to
  {"message":"Hello, World!"} when no dataset file present)
- Bump max_body_bytes 8MB → 32MB and MAX_BODY_BUFFERS 128 → 512
  to accept 20MB uploads
- Fix Dockerfile build context for local testing (USE_LOCAL arg)

+++ src/response/response.zig
@@ -13,7 +13,7 @@ pub const ManagedBody = struct {
 /// to the write queue without linearizing. Used for echo POST optimization.
 pub const ScatteredBody = struct {
     handles: []buffer_pool.BufferHandle,
-    count: u8,
+    count: u16,
     last_buf_len: usize,
     total_len: usize,
     buffer_size: usize,
